requirements.txt
flask
flask-sqlalchemy
supabase
python-dotenv